Made the gradient its own component
